Bug #489294 The local time will overflow after long time bootup
[root cause  ] tv_ptr->sec is a uint32 data, for sec is ok. But when 
switch to us, it will overflow for about just about one and a half hour.
[changes     ] Change the data type to uint64 before switch.
[side effects] No
[self test   ]<9830aef, passed on long time call>:
[reviewers   ] wenmei.hua; Chengli.Du
[change_type ] PLD_Feature_Management
[tag_product ] common

+++ video/video_2_0/jbv/_jbv_packet_loss_utils.c
@@ -123,7 +123,7 @@ void _JBV_checkForPacketLoss(
             obj_ptr->lastMaxSeqn = obj_ptr->lastSeqn;
         }
         /* Find time stamp of current time. */
-        curTime = tv_ptr->sec * _JBV_SEC_TO_USEC + tv_ptr->usec;
+        curTime = ((uint64)tv_ptr->sec) * _JBV_SEC_TO_USEC + (uint64)tv_ptr->usec;
         curTime -= obj_ptr->initTime;
 
         /* Calculate the difference between the incoming packet RTP Seqn and last Rcvd MAXIMUM RTP Seqn. */
